+++ flix/version.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+__version__ = '1.0.0'
+__author__ = 'Chris Griffith'
